requirements.txt
tensorflow-gpu==2.3.1
matplotlib
numpy
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI
tensorflow-addons
tensorflow-datasets
scipy